product mark int -> decimal

+++ init.sql
@@ -35,7 +35,7 @@ CREATE TABLE products (
     `description` TEXT,
     `image` varchar(255),
     amount INT NOT NULL,
-    mark INT DEFAULT 0,
+    mark DECIMAL DEFAULT 0,
     PRIMARY KEY (id)
 );
